Modularize main.rs — extract subcommand handlers into src/cli/ modules
Split the 1,543-line main.rs into 16 focused modules under src/cli/:
one handler per subcommand, plus a feature-gated watch module and a
shared mod.rs with types, enums, dispatch, and helpers. main.rs is
now 8 lines. Pure refactor — no logic changes, all 692 tests pass.

Closes #158
